requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-allauth==65.3.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
WeasyPrint==62.3
gunicorn==23.0.0
whitenoise==6.8.2
